GH#28788: Split collector runtime into quality-safe modules
